Add minspan to sitimechart searches
Over a 5 minute interval, `sitimechart` uses a 5s interval, which is too
many datapoints for a useful summary index. Mostly this applies for the
general resource runtime search; for the by-host search there is not
usually more than one Puppet run in a 5m interval, but I am adding it
for consistency and "just in case".