Split SI catalog compile time searches
Rename the existing "Summary: Catalog Compile Time" by adding "By Node"
to reflect it's purpose. Added a "Total" SI search which runs slightly
less frequently to roll-up the per-node stats and changed the search
used for the compile time chart to use it (and a cron schedule that is
consistent with the summaries).